requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
httpx==0.27.0
langdetect==1.0.9
python-dotenv==1.0.1
pytest==8.3.0
openai>=1.0.0
